digital_twins/structural-monitoring/python/hbk-mqtt-vis/requirements.txt
paho-mqtt==2.1.0
numpy==1.26.4
matplotlib==3.9.0